+++ controllers/app.js
@@ -1,9 +1,22 @@
 const fs = require('fs');
-const path = require('path');
+const showdown = require('showdown');
+
+// showdown converter
+const converter = new showdown.Converter();
 
 
-const palmApi = require('../api/fetchPalm2');
+const palmApi = require('../api/fetchPalm');
+const code = `
+const a = 2;
+const b = 3;
 
+const sum = a + b;
+const product = a * b;
+
+console.log(\`The sum of \${a} and \${b} is \${sum}\`);
+console.log(\`The product of \${a} and \${b} is \${product}\`);
+`
+const description = 'This is a simple code snippet that adds two numbers and prints the sum and product of the two numbers.';
 
 let template;
 
@@ -25,13 +38,25 @@ exports.getApp = (req, res)=>{
 }
 
 exports.postApp = async (req, res)=>{
-    const code = req.body.code;
     console.log('incoming request', req.body);
-    const description = req.body.description;
-    const data = await palmApi.getData2(template, code, description);
+
+    // const code = req.body.code;
+    // const description = req.body.description;
+    
+    const data = await palmApi.getData(template, code, description);
     console.log('data from api', data);
+    
+    fs.writeFile('./data/output.md', data, (err) => {
+        if (err) {
+            console.error(err)
+            return
+        }
+    });
+
+    const html = converter.makeHtml(data);
+
     res.render('output', {
         pageTitle: 'DocAI Generator',
-        data: data
+        data: html
     })
 }
