+++ data/output.md
@@ -1,0 +1,82 @@
+
+# Project Title: Generate Project Title 
+
+## Description
+
+generate Project Description in a multiline paragraph, Explain Project Descriptively
+
+## Table of Contents
+
+- [Installation](#installation)
+- [Usage](#usage)
+- [Code](#code)
+- [Code Explanation](#code-explanation)
+- [Contributing](#contributing)
+- [License](#license)
+
+## Installation
+
+To install and set up this project, follow these steps:
+
+1. Clone the repository to your local machine.
+2. Install the dependencies by running the following command:
+
+```
+npm install
+```
+
+3. Run the project by running the following command:
+
+```
+npm start
+```
+
+## Usage
+
+This project can be used to do the following:
+
+* Generate code snippets.
+* Explain the logic behind code snippets.
+* Provide guidelines for contributing to the project.
+* Provide a license for the project.
+
+## Code
+
+### Code Snippet 1
+
+```
+const a = 2;
+const b = 3;
+
+const sum = a + b;
+const product = a * b;
+
+console.log(`The sum of ${a} and ${b} is ${sum}`);
+console.log(`The product of ${a} and ${b} is ${product}`);
+```
+
+### Code Explanation
+
+This code snippet adds two numbers and prints the sum and product of the two numbers.
+
+The first line of the code snippet declares two variables, `a` and `b`, and assigns them the values 2 and 3, respectively.
+
+The second line of the code snippet declares two variables, `sum` and `product`, and assigns them the values of `a + b` and `a * b`, respectively.
+
+The third and fourth lines of the code snippet print the sum and product of the two numbers to the console.
+
+## Contributing
+
+To contribute to this project, follow these steps:
+
+1. Fork the repository to your own GitHub account.
+2. Create a new branch for your changes.
+3. Make your changes to the code.
+4. Commit your changes to your branch.
+5. Push your changes to your GitHub repository.
+6. Create a pull request to the original repository.
+
+## License
+
+This project is licensed under the MIT License.
+
