+++ package.json
@@ -13,11 +13,12 @@
     "nodemon": "^3.0.1"
   },
   "dependencies": {
+    "@google-ai/generativelanguage": "^1.1.0",
     "axios": "^1.5.1",
     "body-parser": "^1.20.2",
     "ejs": "^3.1.9",
     "express": "^4.18.2",
-    "@google-ai/generativelanguage": "^1.1.0",
-    "google-auth-library": "^9.1.0"
+    "google-auth-library": "^9.1.0",
+    "showdown": "^2.1.0"
   }
 }
