+++ views/output.ejs
@@ -1,5 +1,5 @@
 <%- include('includes/head.ejs') %>
 </head>
 <body>
-        <%= data %><br>
+        <%- data %><br>
 </body>
